+++ frontend/src/app/components/Protectoras/detalle-protectora/detalle-protectora.component.html
@@ -29,8 +29,6 @@
     <button class="btn custom-btn-red-dark p-2 ">Enviar un mensaje</button>    
   </div>
 
-  <div class="p-md-5 d-flex justify-content-center">
-    <button (click)="actualizarProtectora(protectora.idprotectora)" class="btn custom-btn-red-dark p-2 ">Actualizar Protectora</button>        
-  </div>
+
