--- frontend/src/app/components/Protectoras/gestion-protectora/gestion-protectora.component.ts
@@ -1,26 +1,0 @@
-import { Component } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
-import { Protectora } from 'src/app/entidades/protectora';
-import { ProtectoraService } from 'src/app/service/protectora/protectora.service';
-
-@Component({
-  selector: 'app-gestion-protectora',
-  templateUrl: './gestion-protectora.component.html',
-  styleUrls: ['./gestion-protectora.component.css']
-})
-export class GestionProtectoraComponent {
-  id: number;
-  protectora: Protectora;
-  route: ActivatedRoute;
-
-  constructor(private _protectoraService: ProtectoraService, route: ActivatedRoute) {
-    this.route = route;
-  }
-
-  ngOnInit() {
-    this.id = this.route.snapshot.params['id'];
-    this._protectoraService.obtenerProtectoraPorId(this.id).subscribe( (data: Protectora) => {
-      this.protectora = data;
-    });
-  }
-}
