+++ frontend/src/app/components/Protectoras/gestion/gestion-protectora/protectora.component.html
@@ -22,3 +22,4 @@
     </div>
   </div>
 </section>
+
