+++ frontend/src/app/components/Protectoras/lista-protectoras/lista-protectoras.component.html
@@ -15,10 +15,10 @@
         </select>
       </div>
       <div class="col-sm-auto">
-        <button class="custom-btn-red-dark" (click)="listarProtectorasPorIdProvincia(provinciaSeleccionadaId)">Buscar</button>
+        <button class="btn custom-btn-red-dark p-2 " (click)="listarProtectorasPorIdProvincia(provinciaSeleccionadaId)">Buscar</button>
       </div>
       <div class="col-sm-auto">
-        <button class="custom-btn-red-dark" (click)="listar()">Ver todas</button>
+        <button class="btn custom-btn-red-dark p-2 " (click)="listar()">Ver todas</button>
       </div>
     </div>
   </div>
