+++ frontend/src/custom-theme.scss
@@ -39,3 +39,4 @@ button{
   /* Button */
   box-shadow: 0px 2px 10px 0px rgba(101, 192, 237, 0.50);
   }
+
